Assume that the 'overordnet' of the root org. unit points to the org.
Previously, the only way of distinguishing the root organisation unit
from other units was its lack of a superior unit. Unfortunately, we
cannot search for the *absence* of values in LoRA, so the only way of
finding it was to manually inspect all units. This is quite slow with
a dataset of >3000 units...

Instead, we 'extend' the standards and put the UUID of the
organisation itself in the 'overordnet' (aka superior) relation.