basic implementation of effective-date as well as validity
This ensures that correctness of requests with 'validity' set to past
or future; either return a 404 if nothing happens in that time-frame
or return results with 'current' filtered. We can't yield more than
one 'future' or 'past' record, at the moment.

The Time Machine view relies on effective-date for full-history;
unfortunately, that's broken at the moment, so I've marked the test as
an expected failure :(