ensure that we use the correct date & time for the super unit (fixes #3)
I believe the issue reported in #3 was caused by the rename possibly
occurring outside of the validity of the parent. Eventually, our
validation will catch that, but for now, just test that we get it
right when we should.

As a side-effect, this causes us to report incorrect dates for
validity in the parents. We can live with that, given the frontend
doesn't use that data.