Proper SAML authentication support
This time, we add support for obtaining a token from the SAML IdP,
which we attach to all LoRA requests. We stash it in a httpOnly (and
secure, when possible) cookie.

The frontend still contains code for handling authentication, and
still sets the X-AUTH-TOKEN header. We simply disregard it, given that
setting the token in JavaScript is insecure.

I only added tests for WSO2 for now, and they're only unit tests: We
cannot add proper integration or selenium tests without requiring a
proper, running SAML identity provider.