tests: use a temporary file rather than a pipe
Using a pipe can trigger a deadlock of the output from the child
process exceeds a certain limit, and we don't routinely flush/read the
pipe. Originally, we wanted to use a pipe so that we could detect when
the child was ready, and print its output.

We fix this by applying three changes:

1) Use an unnamed temporary file for output from the child.
2) Wait for the child process using the retry support in requests &
   urllib3.
3) Flush the log file regularly, so we retain a reasonably ordering of
   the output in reports from failing tests.