orgunit: change edit API to be consistent with employee
Previously, the API for editing orgunits accepted a `dict`
specifying the orgunit itself. This is both inconsistent with the
employee API, and unsuitable for editing addresses. Instead, we should
accept a list, where each entry specifies a `type` of the relevant
object to edit.

We retain compatibility by transparently wrapping single dict in a
list, and default to the `orgunit` role type.