lora: use proper moment for 'now', if possible, rather than today
This should properly remove the duplicates that we've run into, and
ensure that we only ever display entries that are properly current.

Ideally, we'd use the 'virkningstid' parameter to LoRA for this.
Unfortunately, the current code assumes that overriding 'virkningfra'
and 'virkningtil' in get() and fetch() work quite well. I fixed a few
issues, but the remaining ones were sufficiently annoying that I
simply opted for using a very small interval of one microsecond --
just like LoRA does, in fact.