make parent validity check more robust
This should ensure that a 'candidate parent' is properly valid:

1) It's valid at the exact time of the move.
2) Neither it, nor any parent, belongs to another organisation.
3) We are not creating a cycle.

Test #1 was failing, previously. In fixing I tried to make the
validation routine more robust and readable, adding #2 and
strengthening #3 to check for preëxisting cycles.

The root cause of bug #23182 was that the check didn't use
`common.is_reg_valid()` but instead rolled its own not-quite-as-robust
check, resulting in a move to a unit at an invalid date giving the
wrong error.