common: support lists in get_obj_value()
If we use operator.getitem() and catch LookupError — the superclass of both
KeyError and IndexError — the method matches its intent much more
closely: it's equivalent to performing `x[y]` multiple times.

As a likely unintentional side-effect of its implementation,
get_obj_value() returned an empty dict for incomplete paths, and a
call to `_merge_obj_effects()` happened to rely on this. I fixed that by
adding a specific check for `None`.