auth: login lazily, i.e. when needed
The frontend doesn't know whether it already has an active session, or
even if authentication is active at all, so we shouldn't check the
authentication state prior to loading a page. Instead, simply handle
the error, redirecting to login.

The downside is that the UI briefly flashes when vising a page
directly; the upside is that the latter works at all. In addition,
this ensures that login timeouts work as expected, triggering a login
request, rather than simply breaking the UI.