address: add support for the TEXT scope
We will use this for arbitrary text fields, such as opening hours. In
order to ensure that we round-trip the text accurately, we use a
special quoting for URNs. This is mostly similar to URL quoting,
except that the only characters we consider safe are digits lowercase
letters, digits and '+'.

As an aside, '+' and '%' aren't technically allowed in URNs, but we're
not really using them as URNs anyway, so… ¯\_(ツ)_/¯