change validity end dates to be exclusive
This means that the very last millisecond of a validity period isn't
included -- which almost, but not quite, exactly matches the LoRA
semantics since LoRA uses PostgreSQL time ranges and infinitesimally
excludes the exact timestamp. Irrational numbers are a ever so
slightly obnoxious, so we just use a millisecond instead.

In addition, we clamp the last minute of the day to midnight when
writing.

refs issue 23559