adjust jenkins to the new directory layout
Now that the frontend and backend are largely separate, we use
different script locations for them, each named `.jenkins`. In order
to make the scripts more robust, I extracted shared logic to a
`common.sh` file, and ensured that we clean previous build products
first. (Apparently, Jenkins doesn't actually do so properly...)

As an aside, I figured out how to build the frontend and backend in
parallel, shaving up to a minute off test exectution time.