switch to using dates in REST API
We now use ISO 8601 dates in the API, rather than dates with time.
When reading from LoRA, we simply disregard any date. When validating
requests, anything other than midnight CE[S]T causes a hard error.

This works rather well in practice, although it _may_ lead to issues
should other clients of LoRA write an address or IT system with a
validity changeover not at midnight. In this case, we should be able
to read the data, but not alter it.