date picker: address a major annoyance
Previously, the date picker would always pop up to the current value,
even when that value was disallowed. Instead, clamp whatever we
display to the allowable values.

With our test data, this avoids the need for flipping from somewhere
in the sixties to today in certain circumstances. That's just
obnoxious.