Wait for site to be loaded before running test
Before the selector returned as soon at the elemt was in DOM, but before it was
populated and visible. It cause the .expect() to fail as it was run just after
the selector returned. With the visibilityCheck on the selector, the selector
does not return before the elemet is visible and .expect() is run after
everything is loaded.