[#39855] Don't bubble up error messages to frontend in case of "empty" edits
In `lora.raise_on_status`, detect a specific LoRa error where the
message is on the form:
"ERROR:  Aborted updating bruger with id [...] as the given data, does
not give raise to a new registration. Aborted reg: ...".

In this case, we ignore the error returned by LoRa, and simply return a
null response to the frontend, rather than letting the error message
bubble up to the frontend.

This hopefully also fixes https://redmine.magenta-aps.dk/issues/37453

[#39855] Employee edit modal: handle 'null' MO response, clear error message

This also changes the `cleanUp` method in `MoEmployeeEntry` to be a
public function, so it can be called from its parent
(`MoEntryEditModal`.) This ensures that the unwanted payload fields
("givenname", "surname", "nickname_givenname", "nickname_surname") are
neither displayed nor posted back to MO.

[#39855] `Scope.update`: return original UUID if LoRa indicates a "no-op" change

Update tests accordingly, and add some comments to other test