[#47624] Change return type for queries
This is added to support historic data. Due to the bitemporal nature
of data, there is *not* a 1 to 1 correspondance between UUIDs and
singular models. Rather, a UUID maps to a list of models at different
points in time.

Thus, the response type is changed to a key/value pair of UUIDs and
lists of objects.