[#48310][#48312][#48313][#48314][#48315] "IT Association" in backend/API
An "IT-tilknytning" is defined as an `Association` where the relation
`tilknyttedeitsystemer` contains the UUID of an `ITSystemBinding`, and
the `tilknyttedefunktioner` relation can contain the UUID of a LoRa
class in the facet `engagement_job_function`.

Also, the `ITSystemBinding` itself can now have its `primary` attribute
set.

[#48313] Expose `primary` in IT system bindings

[#48314] Filter associations on "IT system binding"

[#48315] Update reading code to reflect GraphQL requirements; lint

[48312] Support both "normal" and "IT" associations in writing API

[#48312] Fix test breakages due to new fields in API responses

[#48312] Fix test breakage: `validator.is_substitute_allowed` cannot be awaited

[#48310] Incorporate review feedback on strategy pattern and enum

`AssociationReader.get_from_type` uses a strategy pattern to encapsulate
the different paths for reading normal, IT and "substitute"
associations.

`AssociationReader._get_mo_object_from_effect` uses a new enum
`ResponseExtraField` to determine which extra fields to include in the
responses.