[#48310] Proof-of-concept: load and save "IT-tilknytning" objects
An "IT-tilknytning" is defined as an `Association` where the relation
`tilknyttedeitsystemer` contains the UUID of an `ITSystemBinding`, and
the `tilknyttedefunktioner` relation can contain the UUID of a LoRa
class in the facet `engagement_job_function`.

Also, the `ITSystemBinding` itself can now have its `primary` attribute
set.

The code for creating `Association` objects can now create an
`ITSystemBinding` if the required data is available in the request. (The
code uses dummy values for now, where the UUIDs correspond to LoRa
classes available in the AAK dataset.)