[#49089] Introduce feature flag `dar_address_autocomplete_includes_access_addresses`
The feature flag defaults to `True`, preserving the original behavior of
autocompleting on both "normal" addresses *and* "access addresses."

Setting the flag to `False` means that we do not use the DAR
"autocomplete" API to look for "access addresses", but only "normal"
addresses.

This MR contains some PoC code to implement the feature flag. Maybe the
code calling the DAR "autocomplete" API should live in
`os2mo_dar_client` instead?