test: [#xxxxx] remove wierd broken test
The test used to verify that an AttributeError was emitted when an
invalid input was provided to the function.
This seems silly, unless we also test and verify what happens if the
user breaks the function contract in other ways, such as sending None's
from non-nullable arguments, og NaNs for objects, etc.