feat: [#52761] introduce itsystem_create mutator
The mutator is implemented directly atop LoRa's db methods, thereby
bypassing all the writing handlers within OS2mo, and thus the mutator
only works on OS2mo instances with internal LoRa. A guard clause is in
place to ensure that a warning is produced on OS2mo instances running
with an external LoRa instance.

This mutator marks the start of tightly integrating OS2mo and LoRa, and
thus marking the start of OS2mo claiming control over the database.