refactor: [#50647] set database user references via OS2mo middleware
This change restructures how the LoRa database user references on changes
is set. Previously it was hardcoded deep within LoRa, but with this
change it is seeded via a context variable from a middleware in OS2mo.

This change is in itself just a refactoring, but it enables us to modify
the source used for the database user reference in the future.