feat: [#50647] set database user references to keycloak user
This change sets LoRa database user references on changes to the user
uuid set on the keycloak authentification token, such that changes
can be tracked back to the user or integration that made them.
Thereby replacing the current behavior of simply hardcoding user references
to the magical UUID of "42c432e8-9c4a-11e6-9f62-873cf34a735fm".
If no keycloak authentification token can be found, the behavior
defaults back to this magical UUID.

To understand the historic behavior of user references that may be found in
old LoRa databases, please read the timeline below:

May 2015:

User references are introduced into LoRa (e915ea85).
At this time all user references are hardcoded to the UUID:
"615957e8-4aa1-4319-a787-f1f7ad6b5e2c".

August 2015:

Support for wstrust / wso2 is introduced, and if configured for use, the
user reference used is the SAML id.
At the same time a bug is introduced such that the hardcoded uuid from
May of 2015 no longer work when wstrust is disabled, thus leaving broken
references in LoRa when this is the case.

December of 2016:

The bug introduced in August of 2015 is fixed, and a new hardcoded UUID
is introduced, namely: "42c432e8-9c4a-11e6-9f62-873cf34a735fm".
As such user references will no longer be broken, but rather either by
the SAML id or the new magical UUID.

January of 2020:

Support for wstrust / wso2 is removed from LoRa, and thus all user
references are now hardcoded to the magical uuid that was introduced in
December of 2016.

December of 2022:

This pull request is introduced into OS2mo, mapping keycloak user
references to LoRa user references whenever keycloak user references are
available, defaulting to the magical uuid from December of 2016 if it
is not.