chore: [#52084] Copyright without years
The Berne Convention states that copyright “must be automatic; it is
prohibited to require formal registration”. The often-used copyright
lines are not necessary to protect our rights. They can still serve a
purpose as they are informational and make the ownership question quite
clear. The year ranges add questionable value though.[1]

[1] https://daniel.haxx.se/blog/2023/01/08/copyright-without-years/