feat: [#52128] Mark legacy authentication tokens as used
Currently there is no way to see if a legacy authentication token is
alive or not. This has lead to a large number (millions in some cases)
of session tokens being kept alive in the legacy session database.

This change attempts to rectify this by setting expiry to 9999-12-31 for
legacy tokens that are actually being requested, thus enabling us to
eliminate a large number of dead tokens at a later stage.

The long-term goal is still to eliminate the legacy authentication
mechanism entirely, but a short-term goal is to get rid of unused / dead
tokens, and a mid-term goal could be to replace the session database
with environmental variables.