ci: [#42309] Remove keycloak db
This functionality did not work.

With the new Gitlab runners based on Kubernetes, it is not possible to
have two Gitlab CI/CD job services run on the same port. It is possible
on docker. We tried fix it in dca7d1318dfb7ae9f9eaccb560877b36f8b080fd,
by making `mox-db` and `keycloak-db` bind on different ports. It did not
work because the `POSTGRES_PORT` env var is not read. If we used the
correct `PGPORT`, we likely could make it work.

However, Keycloak does not actually use the datatabase. It
prints

> Using Embedded H2 database

on every run.

Instead of having `keycloak-db` start on the correct port and still be
unused, we simply remove it and rely on the embedded database instead.