feat: [54412] `SP_API_VERSION` specifies the API version used for CPR lookups
This anticipates upcoming changes in `service_person_stamdata_udvidet`
which will allow to connect to both version 4 and 5 of the
`/service/CPR/PersonBaseDataExtended` API in Serviceplatformen.

This is necessary to support both existing and new MO customers who use
the CPR lookup facility by way of a service agreement with
Serviceplatformen. Existing customers use version 4 of the API (and
upgrading is tough), while new customers are only able to use version 5,
as Serviceplatformen no longer onboard new customers on version 4.