fix: [#54060] properly filter org unit attributes on edit
prepare_edit would previously get all attribute validities from lora (-inf,inf), and use the last one as the "previous" object in the edit. This is based on the (wrong) assumption that objects do not already have a validity in the future, which is only true if you are editing the latest.
This seems to be true in almost all reading handlers (as they are mostly identical copies of each other). This is tracked separately in #54578.