feat: [#53506] add support for reading registrations
This change exposes bitemporal registrations via our GraphQL interface.
Thus expanding the capability of the GraphQL interface to answer queries,
such as:

* Was a given entity changed in a given time-frame?
* When was a given entity last changed, and who changed it?
* How frequently is a given entity changed?
* What changes have a given user made in a given time-frame?

The change is non-breaking and expands top-level replies with a new
queryable field called `registrations`:
```
query {
  org_units {
    uuid
    registrations {
      registration_id
      start
      end
      actor
    }
  }
}
```
Where a response might look like this:
```
{
  "data": {
    "org_units": [
      {
        "uuid": "1caba8d9-6b9f-506b-b845-9a8c4f5b8a03",
        "registrations": [
          {
            "registration_id": 241,
            "start": "2023-04-19T11:14:35.121135+02:00",
            "end": null,
            "actor": "42c432e8-9c4a-11e6-9f62-873cf34a735f"
          },
          {
            "registration_id": 239,
            "start": "2017-01-12T08:12:24.148346+02:00",
            "end": "2023-04-19T11:14:35.121135+02:00",
            "actor": "42c432e8-9c4a-11e6-9f62-873cf34a735f"
          }
        ]
      }
    ]
  }
}
```
To convey the information that the organisation unit was originally
created on the 12th of January 2017, and then changed on the 19th of
April 2023, to the state that it has now.
Note that registrations do not concern themselves with the historical
changes that occur to the organisation unit itself, only to when changes
to that historical view was made.