feat: [#53354] GraphQL schema extension for LoRa pagination
The "pagination out of range sentinel" - if you will - is implemented by
returning a boolean in the GraphQL extensions field. We experimented
with raising an error, and letting clients check `result.errors`, but
this worked poorly as there seems to be no way to avoid the exception
showing up in the console/log. It seems it is best to express this
expected "error" in the schema[1], but as we will be moving to a
cursor-based approach in a new GraphQL API version shortly (#50653), we
opted for this solution which does not break the existing schema.

[1] https://strawberry.rocks/docs/guides/errors