refactor: [#50474] split up get_detail into multiple functions
This change splits the highly generic:
`/service/{type}/{id}/details/{function}` endpoint, that previously took
type (ou/e) and function (address, association, engagement, etc) and
deconstructs it into 24 specific endpoints.

The purpose of this change is to get endpoint-specific metrics to allow
us to determine which endpoints are actually in use, and which endpoints
are unused and therefore can be removed. Removing endpoints outright is
desirable as it reduces the complexity of the stack and also minimizes
the work required for shimming the Service API.

Additionally this charge also supports the future work of shimming the
Service API, by splitting a highly generic and therefore almost
unshimmable function into very meaningful shimmable handlers.

Finally splitting the endpoint allows FastAPI to generate meaningful and
useful API documentation and OpenAPI Specification.

BREAKING CHANGE:
This change is a breaking change, although very minorly so.

Previously calling with an unknown type or function would return a HTTP
400 status-code, but now it returns a HTTP 404 status-code instead.