fix: [#53506] fix org_unit_refresh lazy resolution
OS2mo 18.21.0 unintentionally broke the `org_unit_refresh` GraphQL
mutator, and by extension via shimming the corresponding
`/ou/{unitid}/refresh` Service API endpoint.

The mutator was broken accidentally by a systematic refactoring
across the entire codebase. This refactoring was okay in all other
instances, however the `org_unit_refresh` mutator is implemented
directly atop the low-level dataloader, and as such was broken when
the low-level dataloader changed its interface in the transition to
lazy resolution of top-level results.

The fact that the mutator was broken without any test-failures from the
test-suite indicates that the mutator is *not* tested rigorously enough
currently.

This change merely fixes the underlying problem *without* actually
introducing more rigorous testing to ensure that a similar regression
occur in the future. A follow-up change should ensure that the mutator
is tested rigorously.

This change *does not* resolve the root-cause of the mutator breaking
either. Namely, this change does not restructure the code such that the
mutator utilizes the top-level resolver instead of the low-level
dataloader. A follow-up change should ensure that the mutator is
refactored to be inline with the rest of the codebase.