feat: [#53506] add support for reading the current object
This change exposes actual-state data via our GraphQL interface.
Thus simplifying the use of our GraphQL interface.

The change is non-breaking and expands top-level replies with a new
queryable field called `current`:
```
query {
  org_units(from_date: null, to_date: null) {
    uuid
    current {
      name
    }
    objects {
      name
      validity {
        from
        to
      }
    }
  }
}
```
Where a response might look like this:
```
{
  "data": {
    "org_units": [
      {
        "uuid": "1caba8d9-6b9f-506b-b845-9a8c4f5b8a03",
        "current": {
          "name": "Egtved børnehus"
        },
        "objects": [
          {
            "name": "Jordrup børnehave",
            "validity": {
              "from": "1960-01-01T00:00:00+01:00",
              "to": "2000-03-31T00:00:00+02:00"
            }
          },
          {
            "name": "Egtved børnehus",
            "validity": {
              "from": "2000-04-01T00:00:00+02:00",
              "to": "2049-12-31T00:00:00+01:00"
            }
          },
          {
            "name": "Kolding Daginstitution",
            "validity": {
              "from": "2050-01-01T00:00:00+01:00",
              "to": null
            }
          }
        ]
      }
    ]
  }
}
```
In objects we see the history of the kindergarten from its inception in
January of 1960, through a journey of increased centralization becoming
Egtved kindergarten and nursery in the 2000, before finally becoming
Kolding daycare center in 2050.

This change simply adds the `current` field which returns the object
entry that is active as of right now. The field might be `null` if none
of the results in `objects` are active at query-time.

It is the goal in the future that the `current` field will always be set
to the query-time value of the entity, irregardless of the values found
in objects. This behavior will be introduced with a future restructuring
of the schema.