fix: [#53506] fix engagement association typed as engagement
OS2mo 18.22.0 unintentionally broke the `engagement_association` field
on the `engagement` entity in the GraphQL interface.

The field was broken accidentally by a mistyping stating that the field
had the type `engagement` rather than `engagement_association`. The
issue was discovered automatically by the hypothesis property-based
testing, although not during the review of the initial merge request.

This fix changes the type of the field to `engagement_association` as it
should have been all along. No further tests are added since the
automated test-suite managed to identify the issue, but a discussion on
the number of examples run though the property-based testing has been
started to address whether more examples should be run, such that
counter-examples and thereby issues are found sooner.