feat: [#53506] mutators return model responses
This change modifies and expands the mutator response type.
Thus expanding the capability of the GraphQL interface.

The change is non-breaking and expands all non-static mutator replies to
respond with `Response`-models alike how the top-level replies behave.

Previously an IT-user could created with a mutator alike this one:
```
mutation {
  ituser_create(
    input: {
      person: "0004b952-a513-430b-b696-8d393d7eb2bb",
      user_key: "Roark",
      itsystem: "a1608e69-c422-404f-a6cc-b873c50af111",
      validity: {from: "2023-01-01"}
    }
  ) {
    uuid
  }
}
```
With `uuid` being the only queryable field on the result, however with
this change we can now query the return as-if we had fetched it using
the top-level field:
```
mutation {
  ituser_create(
    input: {
      person: "0004b952-a513-430b-b696-8d393d7eb2bb",
      user_key: "Roark",
      itsystem: "a1608e69-c422-404f-a6cc-b873c50af111",
      validity: {from: "2023-01-01"}
    }
  ) {
    uuid
    current {
      employee {
        name
      }
    }
  }
}
```
Getting back data about the employee / person we pointed to using the
UUID, by the reference on the resulting object:
```
{
  "data": {
    "ituser_create": {
      "uuid": "adad4ba8-80a5-442c-aa69-a7e818d20577",
      "current": {
        "employee": [
          {
            "name": "Howard Roark"
          }
        ]
      }
    }
  }
}
```