fix(graphql): [#53506] workaround for dangling foreign keys
OS2mo 18.22.0 unintentionally broke the optional list fields in the
GraphQL interface, whenever the underlying fields in the database were
containing dangling foreign key references.

The field was broken accidentally by the assumption that there was no
dangling foreign keys in the database, and that dangling foreign keys
were indeed disallowed by the database itself, however as this
assumption was untrue, the code based on the assumption introduced a
breaking behavior.

This change does not resolve the underlying dangling foreign key issue,
but rather it simply ensures that dangling foreign key references are
handled in a similar way to how they were handled before OS2mo 18.22.0,
such that the unintentional breaking changes is fixed.

A test is also included to document the current behavior in the presence
of dangling foreign keys, and it has been validated that the behavior of
the GraphQL interface is now as it was before OS2mo 18.22.0, or in other
words, that the breaking change only occurs between OS2mo 18.22.0 and
the release containing this commit.

In the future the dangling foreign key reference issue should be
resolved on the database layer, but it is in itself a separate issue
from the resolver reuse introducing a breaking change, and as such, out
of scope for this fix.