docs(changelog): [#53506] expand changelog
This commit changes how the changelog is generated.

Previously the changelog was generated using an upstream Jinja template,
which conservatively did not include commit-messages such as the one
you are reading right now in changelog, however as our application is
open-source the commit-messages are publicly available anyway, and as
such there is no harm in including them in the changelog.

Additionally the changelog now also includes a :warning: icon for
commits that contain breaking changes.

The changelog Jinja template can be found in `docs/changelog.md.j2`.