fix(graphql): [#53506] add missing permission class for mutators
This change adds a previously missing permission class to mutators,
checking that the users are indeed authenticated, before checking
whether they are authorized.

The change has no security implications as the authorization check
implicitly also checks for authenticity as it needs the Keycloak token
roles for the role based access control.

Thus the main purpose of the change is to clean-up the code, while also
improving the usability of the GraphQL interface, by improving the error
message when calling without proper authentication.

Calling without proper authentication before this change, would lead to
the following response:
```
{
  "data": null,
  "errors": [
    {
      "message": "An unknown error occurred.",
      "locations": [
        {
          "line": 1,
          "column": 12
        }
      ],
      "path": [
        "ituser_create"
      ]
    }
  ]
}
```
While with this change, the response becomes:
```
{
  "data": null,
  "errors": [
    {
      "message": "Not authenticated",
      "locations": [
        {
          "line": 1,
          "column": 12
        }
      ],
      "path": [
        "ituser_create"
      ]
    }
  ]
}
```