chore(graphql): [#55433] remove version 1 of GraphQL
BREAKING CHANGE:
This change is a breaking change, although we expect no customers to be
affected, as our metrics report no calls to the GraphQL version in the
last 90 days.

If you do find yourself negatively affected by this change, please
follow the below migration guide to migrate to version 2 of GraphQL.

Assuming a query alike:
```
query OrganisationUnitParentQuery {
  org_units(uuids: [$uuid]) {
    current {
      parent {
        uuid
      }
    }
  }
}
```
The result on version 1 of GraphQL would be:
```
{
  "data": {
    "org_units": [
      {
        "current": {
          "parent": [
            {
              "uuid": "2665d8e0-435b-5bb6-a550-f275692984ef"
            }
          ]
        }
      }
    ]
  }
}
```
While on version 2 of GraphQL the result would be:
```
{
  "data": {
    "org_units": [
      {
        "current": {
          "parent": {
            "uuid": "2665d8e0-435b-5bb6-a550-f275692984ef"
          }
        }
      }
    ]
  }
}
```
The difference is subtle, namely that `parent` used to return a single
element list containing the parent object, while it now returns an
optional parent object instead.

Thus to migrate from version 1 to version 2, simply remove whatever code
that extracts the element from within the list, and use the element
directly.