chore(graphql): [#55433] upgrade strawberry-graphql to 0.151.3
The main purpose of the upgrade is to enable upgrading to a newer
version of FastAPI, as this may be important for security.

Additionally version 0.145.0 of strawberry contains a huge improvement
(according to upstream) with regards to error handling, and thus a huge
boost to developer productivity and happiness.
Please check the strawberry changelog for further details.