fix(graphql): [#55433] fix broken GraphiQL auth and deprecations
OS2mo 20.1.1 upgraded our strawberry-graphql from version 0.137.1 to
0.171.3, and included various changes to our codebase to ensure that the
test-suite kept passing with the newer versions of GraphQL.
However the test coverage seems to have been insufficient as the upgrade
unintentionally broke our custom Javascript injections into GraphiQL,
thus removing the custom Javascript code responsible for deprecation
notices on older GraphQL versions and for redirecting the user to
Keycloak for authentification.

This change introduces a new test-case, which ensures that our custom
Javascript is indeed injected when we expect it to be, and additionally
introduces a change to ensure that the injection does in fact happen
again.

The issue arose from a vaguely documented breaking change in the FastAPI
integrations interface in Strawberry. The Strawberry version that broke
our GraphiQL injections was 0.169.0, and contained the following
breaking change notice:

> Unfortunately, this release does contain some breaking changes, but
> they are minimal and should be quick to fix.
>
> - Flask get_root_value and get_context now receive the request
> - Sanic get_root_value now receives the request and it is async

When reading the changelog it seemed like the breaking changes were the
ones listed, and thus not related to us, as we use the FastAPI
integration, rather than the Flask or Sanic integration, however a
breaking change for FastAPI was also included, although very poorly
documented. As highlighted by Github user kkroening on the related
Pull Request on Github:

> A couple of other subtle breaking changes (AFAICT) with
> `strawberry.fastapi.GraphQLRouter`:
>
> 1. Removal of get_graphiql_response
>
> Use of `GraphQLRouter.get_graphiql_response()` must be changed to
> `GraphQLRouter.render_graphiql(response)`.
> ...

However as this was not part of the changelog, it was not fixed during
the initial upgrade.