ci(dependabot): [#52761] allow 100 dependabot MRs
Before this change dependabot would only create up to 5 simultaneously
open MRs, however only getting 5 when there would potentially be 25,
hides the problem of our old dependencies. We would much rather be
flooded by MRs and know we have a problem, than not be flooded and thus
be unaware of the issue.