feat(auth): [#52129] support legacy sessions via environmental variables
OS2mo used to have a flask-based homebrewed SAML single sign-on solution
implemented using a backing Postgres database. This homebrewed solution
was retired in OS2mo 2.0.0 as Keycloak was added as backing service to
manage authentification and SAML single sign-on.

To ease the transition to Keycloak, the backing database was kept around
along with a new legacy session support system. This system exists till
today and works by checking legacy authentification headers against the
old backing database, allowing access if any of the old SAML-tokens are
used.

As most users have migrated to Keycloak the use of this database has
become mostly obsolete, and only very few session tokens are still in
use.

This change introduces a new environmental variable:

* `os2mo_legacy_sessions`

Which when configured using a JSON-list of UUIDs replace the session
database as the source of legacy session tokens to check against when
using the legacy authentification backend. The goal of introducing this
change is to eliminate the backing sessions database in favor of the new
environmental variable before eliminating the legacy authentification
system in its entirety.