docs(changelog): [#55893] rework old changelog
This change reworks the old changelog such that it is more in line with
the dynamically generated after OS2mo version 12.11.1.

OS2mo has gone through a lot of different release management paradigms,
and by extension a lot of different ways to generate our changelog.

Roughly 4 different paradigms have been in play:

* Before version 0.9.0

At this stage release management was an entirely manual process and no
changelog was being maintained, as such no changelog exists from OS2mo
releases prior to version 0.9.0 and the best place to look for a
changelog is probably the commit history, although no convention was in
place for that either, so the commit history at this change is of
doubtful quality.

* From version 0.9.0 to version 1.18.1

At this stage release management was still an entirely manual process,
however a changelog was manually kept and maintained, and as such all
versions from 0.9.0 have entries in the changelog.

* From version 2.0.0 to version 12.11.1

At this stage release management became systematized by utilizing
[autopub](https://github.com/autopub/autopub) as our release management
software, automation and convention, and as such all releases from
version 2.0.0 have entries in the changelog albeit with a different
format from version before version 2.0.0.

This change was motivated by the need for release automation and the
prospect of removing the release manager role from our team.

* After version 12.11.1

At this stage release management was changed from autopub to the vastly
more popular [semantic-release](https://github.com/semantic-release/semantic-release),
with our changelog being generated by [commitizen](https://github.com/commitizen-tools/commitizen).

This change was motivated by our transition to conventional commits as
our commit message convention, in part itself motivated by the much
better off-the-shelf tooling available using a known convention.
Additionally we were motivated by the prospect of enabling merge-trains
to ease and improve developer productivity by side-stepping the issue of
continuously having to rebase ones merge-request in an attempt to get it
merged.